src/main/java/com/java/demos/mockito/service/AuthenticationService.java
package com.java.demos.mockito.service;

import com.java.demos.mockito.User;
import com.java.demos.mockito.data.UserRepository;

public class AuthenticationService {
    private UserRepository userRepository;

    public AuthenticationService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public boolean authenticate(String username, String password) {
        User user = userRepository.findByUsername(username);
        return user.getPassword().equals(password);
    }
}
